+++ packages/backend/src/main.ts
@@ -1,4 +1,4 @@
-import { readFile } from 'fs/promises';
+import { readFile, rm } from 'fs/promises';
 import { existsSync, watch } from 'fs';
 import { SourcebotConfigurationSchema } from "./schemas/v2.js";
 import { getGitHubReposFromConfig } from "./github.js";
@@ -15,6 +15,8 @@ import stripJsonComments from 'strip-json-comments';
 import { indexGitRepository, indexLocalRepository } from "./zoekt.js";
 import { getLocalRepoFromConfig, initLocalRepoFileWatchers } from "./local.js";
 import { captureEvent } from "./posthog.js";
+import { glob } from 'glob';
+import path from 'path';
 
 const logger = createLogger('main');
 
@@ -65,6 +67,67 @@ const syncLocalRepository = async (repo: LocalRepository, settings: Settings, ct
     return {
         indexDuration_s,
     }
+}
+
+export const deleteStaleRepository = async (repo: Repository, db: Database, ctx: AppContext) => {
+    logger.info(`Deleting stale repository ${repo.id}:`);
+
+    // Delete the checked out git repository (if applicable)
+    if (repo.vcs === "git") {
+        logger.info(`\tDeleting git directory ${repo.path}...`);
+        await rm(repo.path, {
+            recursive: true
+        });
+    }
+
+    // Delete all .zoekt index files
+    {
+        // .zoekt index files are named with the repository name,
+        // index version, and shard number. Some examples:
+        //
+        //   git repos:
+        //   github.com%2Fsourcebot-dev%2Fsourcebot_v16.00000.zoekt
+        //   gitlab.com%2Fmy-org%2Fmy-project.00000.zoekt
+        //
+        //   local repos:
+        //   UnrealEngine_v16.00000.zoekt
+        //   UnrealEngine_v16.00001.zoekt
+        //   ...
+        //   UnrealEngine_v16.00016.zoekt
+        //
+        // Notice that local repos are named with the repository basename and
+        // git repos are named with the query-encoded repository name. Form a
+        // glob pattern with the correct prefix & suffix to match the correct
+        // index file(s) for the repository.
+        //
+        // @see : https://github.com/sourcegraph/zoekt/blob/c03b77fbf18b76904c0e061f10f46597eedd7b14/build/builder.go#L348
+        const indexFilesGlobPattern = (() => {
+            switch (repo.vcs) {
+                case 'git':
+                    return `${encodeURIComponent(repo.id)}*.zoekt`;
+                case 'local':
+                    return `${path.basename(repo.path)}*.zoekt`;
+            }
+        })();
+
+        const indexFiles = await glob(indexFilesGlobPattern, {
+            cwd: ctx.indexPath,
+            absolute: true
+        });
+
+        await Promise.all(indexFiles.map((file) => {
+            logger.info(`\tDeleting index file ${file}...`);
+            return rm(file);
+        }));
+    }
+
+    // Delete db entry
+    logger.info(`\tDeleting db entry...`);
+    await db.update(({ repos }) => {
+        delete repos[repo.id];
+    });
+    
+    logger.info(`Deleted stale repository ${repo.id}`);
 }
 
 /**
@@ -137,6 +200,7 @@ const syncConfig = async (configPath: string, db: Database, signal: AbortSignal,
     // Update the settings
     const updatedSettings: Settings = {
         maxFileSize: config.settings?.maxFileSize ?? DEFAULT_SETTINGS.maxFileSize,
+        autoDeleteStaleRepos: config.settings?.autoDeleteStaleRepos ?? DEFAULT_SETTINGS.autoDeleteStaleRepos,
     }
     const _isAllRepoReindexingRequired = isAllRepoReindexingRequired(db.data.settings, updatedSettings);
     await updateSettings(updatedSettings, db);
@@ -292,10 +356,16 @@ export const main = async (context: AppContext) => {
         for (const [_, repo] of Object.entries(repos)) {
             const lastIndexed = repo.lastIndexedDate ? new Date(repo.lastIndexedDate) : new Date(0);
 
-            if (
-                repo.isStale ||
-                lastIndexed.getTime() > Date.now() - REINDEX_INTERVAL_MS
-            ) {
+            if (repo.isStale) {
+                if (db.data.settings.autoDeleteStaleRepos) {
+                    await deleteStaleRepository(repo, db, context);
+                } else {
+                    // skip deletion...
+                }
+                continue;
+            }
+
+            if (lastIndexed.getTime() > Date.now() - REINDEX_INTERVAL_MS) {
                 continue;
             }